+++ storage/ndb/test/ndbapi/testBasic.cpp
@@ -1125,7 +1125,7 @@ runTupErrors(NDBT_Context* ctx, NDBT_Step* step){
   const NdbDictionary::Table * tab = ctx->getTab();
   Uint32 i;
   int bits = TupError::TE_MULTI_OP;
-  for(i = 0; i<tab->getNoOfColumns(); i++)
+  for(i = 0; i<(Uint32)tab->getNoOfColumns(); i++)
   {
     if (tab->getColumn(i)->getArrayType() != NdbDictionary::Column::ArrayTypeFixed)
       bits |= TupError::TE_VARSIZE;
@@ -1452,7 +1452,7 @@ runBug20535(NDBT_Context* ctx, NDBT_Step* step)
   NdbDictionary::Dictionary * dict = pNdb->getDictionary();
 
   bool null = false;
-  for (i = 0; i<tab->getNoOfColumns(); i++)
+  for (i = 0; i<(Uint32)tab->getNoOfColumns(); i++)
   {
     if (tab->getColumn(i)->getNullable())
     {
@@ -1477,7 +1477,7 @@ runBug20535(NDBT_Context* ctx, NDBT_Step* step)
   pOp = pTrans->getNdbOperation(tab->getName());
   pOp->insertTuple();
   hugoTrans.equalForRow(pOp, 0);
-  for (i = 0; i<tab->getNoOfColumns(); i++)
+  for (i = 0; i<(Uint32)tab->getNoOfColumns(); i++)
   {
     if (!tab->getColumn(i)->getPrimaryKey() &&
         !tab->getColumn(i)->getNullable())
@@ -1496,7 +1496,7 @@ runBug20535(NDBT_Context* ctx, NDBT_Step* step)
   pOp->readTuple();
   hugoTrans.equalForRow(pOp, 0);
   Vector<NdbRecAttr*> values;
-  for (i = 0; i<tab->getNoOfColumns(); i++)
+  for (i = 0; i<(Uint32)tab->getNoOfColumns(); i++)
   {
     if (!tab->getColumn(i)->getPrimaryKey() &&
         tab->getColumn(i)->getNullable())
